requirements.txt
fastapi
uvicorn
pillow
moviepy
requests
python-multipart
